Optical redesign of compound creator dialog
New design of compound creator dialog to tabs for every type of
property. This will allow future additions of new property types (like
electrolyte data) in a nice and easy way.

Eliminated bug: saved files are not loaded as always a regression study
was created. Also *.dwxml files were not displayed as file type for
loading.